requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.7
django-allauth==65.5.0
django-countries==7.6.1
django-crispy-forms==2.3
gunicorn==23.0.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
